Struct indicatif::ProgressDrawTarget

pub struct ProgressDrawTarget { /* fields omitted */ }

Target for draw operations

This tells a progress bar or a multi progress object where to paint to. The draw target is a stateful wrapper over a drawing destination and internally optimizes how often the state is painted to the output device.

Methods

impl ProgressDrawTarget

fn stdout() -> ProgressDrawTarget

Draw to a buffered stdout terminal at a max of 15 times a second.

This is the default draw target for progress bars. For more information see ProgressDrawTarget::to_term.

fn stderr() -> ProgressDrawTarget

Draw to a buffered stderr terminal at a max of 15 times a second.

For more information see ProgressDrawTarget::to_term.

fn to_term(term: Term, refresh_rate: Option<u64>) -> ProgressDrawTarget

Draw to a terminal, optionally with a specific refresh rate.

Progress bars are by default drawn to terminals however if the terminal is not user attended the entire progress bar will be hidden. This is done so that piping to a file will not produce useless escape codes in that file.

fn hidden() -> ProgressDrawTarget

A hidden draw target.

This forces a progress bar to be not rendered at all.

fn is_hidden(&self) -> bool

Returns true if the draw target is hidden.

This is internally used in progress bars to figure out if overhead from drawing can be prevented.