Struct indicatif::MultiProgress

pub struct MultiProgress { /* fields omitted */ }

Manages multiple progress bars from different threads.

Methods

impl MultiProgress

fn new() -> MultiProgress

Creates a new multi progress object that draws to stdout.

fn set_draw_target(&self, target: ProgressDrawTarget)

Sets a different draw target for the multiprogress bar.

fn add(&self, bar: ProgressBar) -> ProgressBar

Adds a progress bar.

The progress bar added will have the draw target changed to a remote draw target that is intercepted by the multi progress object.

fn join(&self) -> Result<()>

Waits for all progress bars to report that they are finished.

You need to call this as this will request the draw instructions from the remote progress bars. Not calling this will deadlock your program.

fn join_and_clear(&self) -> Result<()>

Works like join but clears the progress bar in the end.