Struct indicatif::ProgressBar

pub struct ProgressBar { /* fields omitted */ }

A progress bar or spinner.

Methods

impl ProgressBar

fn new(len: u64) -> ProgressBar

Creates a new progress bar with a given length.

This progress bar by default draws directly to stdout.

fn hidden() -> ProgressBar

Creates a completely hidden progress bar.

This progress bar still responds to API changes but it does not have a length or render in any way.

fn new_spinner() -> ProgressBar

Creates a new spinner.

This spinner by default draws directly to stdout. This adds the default spinner style to it.

fn set_style(&self, style: ProgressStyle)

Overrides the stored style.

fn tick(&self)

Manually ticks the spinner or progress bar.

This automatically happens on any other change to a progress bar.

fn inc(&self, delta: u64)

Advances the position of a progress bar by delta.

fn set_position(&self, pos: u64)

Sets the position of the progress bar.

fn set_length(&self, len: u64)

Sets the length of the progress bar.

fn set_prefix(&self, prefix: &str)

Sets the current prefix of the progress bar.

fn set_message(&self, msg: &str)

Sets the current message of the progress bar.

fn finish(&self)

Finishes the progress bar and leaves the current message.

fn finish_with_message(&self, msg: &str)

Finishes the progress bar and sets a message.

fn finish_and_clear(&self)

Finishes the progress bar and completely clears it.

fn set_draw_target(&self, target: ProgressDrawTarget)

Sets a different draw target for the progress bar.

This can be used to draw the progress bar to stderr for instance:

let pb = ProgressBar::new(100);
pb.set_draw_target(ProgressDrawTarget::stderr());