Trait indicatif::ProgressIterator

pub trait ProgressIteratorwhere
    Self: Sized + Iterator,{
    fn progress_with(self, progress: ProgressBar) -> ProgressBarIter<Self> ;

    fn try_progress(self) -> Option<ProgressBarIter<Self>> { ... }
    fn progress(self) -> ProgressBarIter<Self> 
    where
        Self: ExactSizeIterator,
    { ... }
    fn progress_count(self, len: u64) -> ProgressBarIter<Self>  { ... }
    fn progress_with_style(self, style: ProgressStyle) -> ProgressBarIter<Self> 
    where
        Self: ExactSizeIterator,
    { ... }
}

Wraps an iterator to display its progress.

Required Methods

fn progress_with(self, progress: ProgressBar) -> ProgressBarIter<Self>

Wrap an iterator with a custom progress bar.

Provided Methods

fn try_progress(self) -> Option<ProgressBarIter<Self>>

Wrap an iterator with default styling. Uses Iterator::size_hint to get length. Returns Some(..) only if size_hint.1 is Some. If you want to create a progress bar even if size_hint.1 returns None use progress_count or progress_with instead.

fn progress(self) -> ProgressBarIter<Self> where
Self: ExactSizeIterator,

Wrap an iterator with default styling.

fn progress_count(self, len: u64) -> ProgressBarIter<Self>

Wrap an iterator with an explicit element count.

fn progress_with_style(self, style: ProgressStyle) -> ProgressBarIter<Self> where
Self: ExactSizeIterator,

Wrap an iterator with a progress bar and style it.